requirements.txt
streamlit>=1.28.0
openai>=1.0.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0

# Speaker Diarization (optional)
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
